Refactor: Move random Milty setup logic to MiltyRandomSetupService (#5302)
Move random setup methods from MiltyService to MiltyRandomSetupService
